requirements.txt
glfw>=2.5.3
PyOpenGL>=3.1.6
numpy>=1.21.0
